requirements.txt
paho-mqtt
pandas
plotly